Created Gravitas project with Intel XDK
This commit includes a playable character with a simple map.

To start editing this project, you can clone and put this project into
www directory. Once it is inside, you should be able to execute the
game. I have corrected all the path directories in stage.html. You
should be able to run it normally once this folder is in your www
directory.